chore(line-chart-native): rename loading series file
